pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drf-perm-control"
version = "0.1.0"
description = "A flexible Django REST Framework permission class with Redis caching support"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Archie", email = "me@archieyang.com"}
]
keywords = ["django", "drf", "permissions", "rest-framework", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
requires-python = ">=3.9"
dependencies = [
    "Django>=4.0",
    "djangorestframework>=3.14",
]

[project.optional-dependencies]
redis = ["redis>=4.0", "django-redis>=5.0"]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/Archie818/drf-perm-control"
Documentation = "https://github.com/Archie818/drf-perm-control#readme"
Repository = "https://github.com/Archie818/drf-perm-control.git"
Issues = "https://github.com/Archie818/drf-perm-control/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

